format and separate ServiceFunction
